create TokenAmountForm component and use it in Mint
